requirements.txt
black==22.3.0
click==8.1.3
colorama==0.4.4
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
mysql-client==0.0.1
mysql-connector-python==8.0.29
mysqlclient==2.1.0
pathspec==0.9.0
platformdirs==2.5.2
protobuf==3.20.1
python-dotenv==0.20.0
SQLAlchemy==1.4.36
tomli==2.0.1
Werkzeug==2.1.2
